master2/METHODOLOGIE_TESTS/6_TP/methodologie-du-test/session_4/requirements.txt
pytest
pytest-cov
python-dotenv
ruff
bs4
requests
python-slugify
unicode
mypy
pandas
scikit-learn
hrid
great_expectations
numpy
